% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supercut}
\alias{supercut}
\title{(Better) Convert Numeric to Factor}
\usage{
supercut(x, intervals, labels = NULL, ordered_result = T)
}
\arguments{
\item{x}{a numeric vector which is to be converted to a factor by supercutting}

\item{intervals}{a vector of string intervals, see Details and Examples}

\item{labels}{labels (if not given names(intervals) will be used if set)}

\item{ordered_result}{logical: should the result be an ordered factor?}
}
\description{
Like \code{cut}, \code{supercut} divides the range of x into intervals and
codes the values in x according to the interval in which they fall. Unlike
\code{cut}, it understands mathematical interval notation so you can be very
precise with how the intervals are constructed.
}
\details{
The intervals are strings being either a single number (e.g. "10"), which
will be assigned to only that value, or an interval (e.g. "[10,20]"), which
will be assigned to any value in that interval. Open ("(1,5)"), closed
("[1,5]") and mixed ("[1,5)") intervals are supported. To create an interval
which is open below, use e.g. "(-Inf, 5]" and for an interval which is open
above use e.g. "(20, Inf)".

If a value in \code{x} doesn't match any interval, a warning will be issued.

If intervals overlap behaviour is unspecified (don't do this!)
}
\examples{
supercut(mtcars$hp, c(
  "less than 110 hp" = "(-Inf,110)",
  "exactly 110 hp"   = "110",
  "110 - 200 hp"     = "(110,200)",
  "200 hp and over"  = "[200, Inf)"
))

}
\seealso{
\code{\link{cut}}
}
