% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbggeo.R
\docType{package}
\name{wbggeo-package}
\alias{wbggeo}
\alias{wbggeo-package}
\title{wbggeo: Geospatial dataviz for World Bank Group}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Andrew Whitby \email{awhitby@worldbank.org}

}
\keyword{internal}
